+++ app/controllers/modern_queue_dashboard/queues_controller.rb
@@ -10,10 +10,12 @@ module ModernQueueDashboard
       @queue_name = params[:id]
       @queue = QueueSummary.with_stats.detect { |q| q.name == @queue_name }
 
-      return if @queue
-
-      flash[:error] = "Queue not found"
-      redirect_to queues_path
+      if @queue
+        @jobs = JobSummary.for_queue(@queue_name, 50)
+      else
+        flash[:error] = "Queue not found"
+        redirect_to queues_path
+      end
     end
   end
 end